normalize score by number of identifiers

+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/domain/score/GeometricDescriptorScore.java
@@ -5,12 +5,12 @@ public class GeometricDescriptorScore implements Score {
     private static final double SIDE_CHAIN_WEIGHT = 1;
     private static final double ANGLE_WEIGHT = 1;
 
-    private final int backboneScore;
-    private final int sideChainScore;
-    private final int angleScore;
+    private final double backboneScore;
+    private final double sideChainScore;
+    private final double angleScore;
     private final double score;
 
-    public GeometricDescriptorScore(int backboneScore, int sideChainScore, int angleScore) {
+    public GeometricDescriptorScore(double backboneScore, double sideChainScore, double angleScore) {
         this.backboneScore = backboneScore;
         this.sideChainScore = sideChainScore;
         this.angleScore = angleScore;
@@ -22,15 +22,15 @@ public class GeometricDescriptorScore implements Score {
         return score;
     }
 
-    public int getBackboneScore() {
+    public double getBackboneScore() {
         return backboneScore;
     }
 
-    public int getSideChainScore() {
+    public double getSideChainScore() {
         return sideChainScore;
     }
 
-    public int getAngleScore() {
+    public double getAngleScore() {
         return angleScore;
     }
 }
